Sample Practice test file
Testing with Files and directories in JUnit is easy thanks to TemporaryFolder @Rule. 
In JUnit rules (@Rule) can be used as an alternative or an addition to fixture setup and 
cleanup methods (org.junit.Before, org.junit.After, org.junit.BeforeClass, and org.junit.AfterClass), but they are more powerful, and can be more easily shared between projects and classes.

The TemporaryFolder rule provides two methods to manage files and directories: newFile and newFolder. Both method return desired object under the temporary folder created in the setup method. In case the path of the temporary folder itself is needed getRoot method of TemporaryFolder can be used.

Everything that will be added to the temp folder during tests will be automatically removed when test finishes, despite it was successful or not.